app/requirements.txt
django==3.2
djangorestframework
drf-yasg
django-cors-header
psycopg2-binary
gunicorn
telepot